erm
Some bug fixing on my side , fixing the translation issues I had , played around with the voice input but I didn't manage to get it to work completely so I just won't keep it in this update. Tts speaks out just fine which is a dub. Audio input grind next